fix(fe): default notif toggle on and attach install listener at module load
Two demo-flow fixes:

1. Pre-tick: notifications toggle now defaults to on for first-time
   visitors. Previously defaulted to off, so users had to manually flip
   it. Server-confirmed "not consented" no longer resets the toggle
   back to off — only server-confirmed "consented" pins it on.

2. Install prompt timing: extract beforeinstallprompt capture into
   $lib/utils/pwaInstall so the listener is attached at module load
   time. The previous in-onMount attach was too late — Chrome fires the
   event during initial page load and the ContinueView component
   wasn't yet mounted, so the deferred prompt was lost. Module-scope
   attach runs when the util is first imported (via ContinueView's
   import chain), which happens synchronously during page setup.
